% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExerG.R
\name{ExerG}
\alias{ExerG}
\title{R DATA ANALYSIS}
\usage{
ExerG(x)
}
\arguments{
\item{x}{The exercise number (ranging from 1 to 6).}
}
\value{
The solution for the specified exercise.
}
\description{
In this work, an analysis and different exercises are carried out with data from the nycflights13 package, using additional libraries such as Tidyverse, knitr, among other additional functions
}
\examples{
# Retrieve the solution for exercise 1
ExerG1

# Retrieve the solution for exercise 6
ExerG6
}
